refactor: normalize projects and split product workflows
Implemented the semantic/session product normalization and toolchain convergence campaign end to end.

- normalized durable project/repo semantics through one canonical layer
- split session-product lifecycle, archive-product operator, semantic surface, schema tooling, and parser seams by role
- added archive-debt governance stages plus semantic-product validation lanes
- rebuilt live session products and exercised the live archive through the new lane
