+++ devtools/run_validation_lanes.py
@@ -126,6 +126,33 @@ LANES: dict[str, LaneConfig] = {
             "tests/integration/test_health.py",
         ],
     ),
+    "semantic-product-normalization": LaneConfig(
+        name="semantic-product-normalization",
+        description="Semantic/session product normalization, operator/toolchain narrowing, schema contracts, and provider parser cleanup",
+        timeout_s=900,
+        command=[
+            sys.executable,
+            "-m",
+            "pytest",
+            "-q",
+            "-n",
+            "0",
+            "tests/unit/core/test_project_normalization.py",
+            "tests/unit/cli/test_products.py",
+            "tests/unit/core/test_facade_api.py",
+            "tests/unit/mcp/test_tool_contracts.py",
+            "tests/unit/cli/test_check.py",
+            "tests/unit/core/test_schema_registry.py",
+            "tests/unit/core/test_schema_generation.py",
+            "tests/unit/core/test_operator_models.py",
+            "tests/integration/test_schema_operator_workflow.py",
+            "tests/integration/test_schema_evidence_roundtrip_lane.py",
+            "tests/unit/storage/test_search_misc.py",
+            "tests/unit/sources/test_parsers_base.py",
+            "tests/unit/sources/test_parsers_drive.py",
+            "tests/unit/devtools/test_validation_lanes.py",
+        ],
+    ),
     "retrieval-dogfood": LaneConfig(
         name="retrieval-dogfood",
         description="Action-aware query truth, grouped retrieval stats, archive health, and MCP retrieval payload coverage",
@@ -291,6 +318,26 @@ LANES: dict[str, LaneConfig] = {
             "tags",
             "--limit",
             "20",
+            "--json",
+        ],
+    ),
+    "live-products-day-summaries": LaneConfig(
+        name="live-products-day-summaries",
+        description="Live archive day-summary product surface over the recent semantic slice",
+        timeout_s=180,
+        command=[
+            sys.executable,
+            "-m",
+            "polylogue",
+            "--plain",
+            "products",
+            "day-summaries",
+            "--provider",
+            "claude-code",
+            "--since",
+            "2026-03-01",
+            "--limit",
+            "14",
             "--json",
         ],
     ),
@@ -512,6 +559,24 @@ LANES: dict[str, LaneConfig] = {
         description="Full runtime-substrate validation lane covering local contracts plus bounded live archive checks",
         timeout_s=3600,
         sub_lanes=("runtime-substrate-contracts", "runtime-substrate-live"),
+    ),
+    "semantic-product-live": LaneConfig(
+        name="semantic-product-live",
+        description="Bounded live archive lane for normalized product surfaces, governance preview, and maintenance budgets",
+        timeout_s=1800,
+        sub_lanes=(
+            "live-products-status",
+            "live-products-tags",
+            "live-products-day-summaries",
+            "live-products-debt",
+            "live-governance-small",
+        ),
+    ),
+    "semantic-product-hardening": LaneConfig(
+        name="semantic-product-hardening",
+        description="Full semantic-product normalization and toolchain convergence lane",
+        timeout_s=3600,
+        sub_lanes=("semantic-product-normalization", "semantic-product-live"),
     ),
     "frontier-local": LaneConfig(
         name="frontier-local",
