+++ polylogue/schemas/generation_workflow.py
@@ -7,29 +7,27 @@ from pathlib import Path
 from typing import Any
 
 from polylogue.paths import db_path as default_db_path
-from polylogue.schemas.field_stats import _collect_field_stats
 from polylogue.schemas.generation_analysis import (
     _artifact_priority,
-    _build_package_candidates,
     _cluster_profile_tokens,
     _cluster_reservoir_size,
     _cluster_sort_key,
     _collect_cluster_accumulators,
+)
+from polylogue.schemas.generation_models import GenerationResult, _ProviderBundle, _UnitMembership
+from polylogue.schemas.generation_packages import (
+    _build_package_candidates,
     _element_profile_tokens,
     _membership_observed_window,
     _merge_representative_paths,
 )
-from polylogue.schemas.generation_models import GenerationResult, _ProviderBundle, _UnitMembership
-from polylogue.schemas.generation_support import (
-    GENSON_AVAILABLE,
-    SchemaBuilder,
-    _annotate_schema,
-    _annotate_semantic_and_relational,
-    _build_redaction_report,
-    _remove_nested_required,
-    collapse_dynamic_keys,
+from polylogue.schemas.generation_schema_builder import (
+    _apply_schema_metadata,
+    _generate_cluster_schema,
+    generate_schema_from_samples,
 )
-from polylogue.schemas.observation import PROVIDERS, ProviderConfig, resolve_provider_config
+from polylogue.schemas.generation_support import GENSON_AVAILABLE
+from polylogue.schemas.observation import PROVIDERS, resolve_provider_config
 from polylogue.schemas.packages import (
     SchemaElementManifest,
     SchemaPackageCatalog,
@@ -37,124 +35,7 @@ from polylogue.schemas.packages import (
 )
 from polylogue.schemas.redaction_report import SchemaReport
 from polylogue.schemas.registry import ClusterManifest, SchemaCluster, SchemaRegistry
-from polylogue.schemas.shape_fingerprint import _structure_fingerprint
 from polylogue.types import Provider
-
-_STRUCTURE_EXEMPLARS_PER_FINGERPRINT = 8
-
-
-def _generate_cluster_schema(
-    provider: str,
-    config: ProviderConfig,
-    samples: list[dict[str, Any]],
-    conv_ids: list[str | None],
-    *,
-    privacy_config: Any | None,
-) -> tuple[dict[str, Any], SchemaReport | None]:
-    """Generate one schema version from the bounded cluster reservoir."""
-    if not samples:
-        return {"type": "object", "description": "No samples available"}, None
-
-    builder = SchemaBuilder()
-    fingerprint_counts: dict[Any, int] = {}
-    for sample in samples:
-        fingerprint = _structure_fingerprint(sample)
-        seen = fingerprint_counts.get(fingerprint, 0)
-        if seen < _STRUCTURE_EXEMPLARS_PER_FINGERPRINT:
-            builder.add_object(sample)
-            fingerprint_counts[fingerprint] = seen + 1
-
-    schema = builder.to_schema()
-    schema = collapse_dynamic_keys(schema)
-    schema = _remove_nested_required(schema)
-    if config.sample_granularity == "record":
-        schema.pop("required", None)
-
-    conv_ids_for_stats: list[str | None] | None = conv_ids if any(conv_id is not None for conv_id in conv_ids) else None
-    field_stats = _collect_field_stats(samples, conversation_ids=conv_ids_for_stats)
-    schema = _annotate_schema(
-        schema,
-        field_stats,
-        min_conversation_count=3,
-        privacy_config=privacy_config,
-    )
-    schema = _annotate_semantic_and_relational(schema, field_stats)
-    schema["$schema"] = "https://json-schema.org/draft/2020-12/schema"
-
-    redaction_report = _build_redaction_report(
-        provider,
-        field_stats,
-        schema,
-        privacy_config=privacy_config,
-        privacy_level=getattr(privacy_config, "level", "standard") if privacy_config else "standard",
-    )
-    return schema, redaction_report
-
-
-def _apply_schema_metadata(
-    schema: dict[str, Any],
-    *,
-    provider: str,
-    config: ProviderConfig,
-    schema_sample_count: int,
-    anchor_profile_family_id: str,
-    artifact_kind: str,
-    observed_artifact_count: int,
-) -> None:
-    schema["title"] = f"{provider} export format ({artifact_kind})"
-    schema["description"] = config.description
-    schema["x-polylogue-generated-at"] = datetime.now(tz=timezone.utc).isoformat()
-    schema["x-polylogue-sample-count"] = schema_sample_count
-    schema["x-polylogue-generator"] = "polylogue.schemas.schema_inference"
-    schema["x-polylogue-sample-granularity"] = config.sample_granularity
-    schema["x-polylogue-anchor-profile-family-id"] = anchor_profile_family_id
-    schema["x-polylogue-observed-artifact-count"] = observed_artifact_count
-    schema["x-polylogue-artifact-kind"] = artifact_kind
-
-
-def generate_schema_from_samples(
-    samples: list[dict[str, Any]],
-    *,
-    annotate: bool = True,
-    max_stats_samples: int = 500,
-    max_genson_samples: int | None = None,
-) -> dict[str, Any]:
-    """Generate JSON schema from samples using genson, with optional annotations."""
-    if not GENSON_AVAILABLE:
-        raise ImportError("genson is required for schema generation. Install with: pip install genson")
-
-    if not samples:
-        return {"type": "object", "description": "No samples available"}
-
-    genson_samples = samples
-    if max_genson_samples and len(samples) > max_genson_samples:
-        import random
-
-        rng = random.Random(0)
-        genson_samples = rng.sample(samples, max_genson_samples)
-
-    builder = SchemaBuilder()
-    for sample in genson_samples:
-        builder.add_object(sample)
-
-    schema = builder.to_schema()
-    schema = collapse_dynamic_keys(schema)
-    schema = _remove_nested_required(schema)
-
-    if annotate:
-        stats_samples = samples
-        if max_stats_samples and len(samples) > max_stats_samples:
-            import random
-
-            rng = random.Random(42)
-            stats_samples = rng.sample(samples, max_stats_samples)
-
-        field_stats = _collect_field_stats(stats_samples)
-        schema = _annotate_schema(schema, field_stats)
-        schema = _annotate_semantic_and_relational(schema, field_stats)
-
-    schema["$schema"] = "https://json-schema.org/draft/2020-12/schema"
-    return schema
 
 
 def generate_provider_schema(
