+++ polylogue/storage/backends/queries/maintenance_runs.py
@@ -2,6 +2,8 @@
 
 from __future__ import annotations
 
+import sqlite3
+
 import aiosqlite
 
 from polylogue.storage.backends.queries.mappers import _row_to_maintenance_run_record
@@ -10,6 +12,7 @@ from polylogue.storage.store import MaintenanceRunRecord, _json_array_or_none, _
 __all__ = [
     "list_maintenance_runs",
     "record_maintenance_run",
+    "record_maintenance_run_sync",
 ]
 
 
@@ -50,6 +53,43 @@ async def record_maintenance_run(
     )
     if transaction_depth == 0:
         await conn.commit()
+
+
+def record_maintenance_run_sync(
+    conn: sqlite3.Connection,
+    record: MaintenanceRunRecord,
+) -> None:
+    conn.execute(
+        """
+        INSERT INTO maintenance_runs (
+            maintenance_run_id,
+            schema_version,
+            executed_at,
+            mode,
+            preview,
+            repair_selected,
+            cleanup_selected,
+            vacuum_requested,
+            target_names_json,
+            success,
+            manifest_json
+        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
+        """,
+        (
+            record.maintenance_run_id,
+            record.schema_version,
+            record.executed_at,
+            record.mode,
+            int(record.preview),
+            int(record.repair_selected),
+            int(record.cleanup_selected),
+            int(record.vacuum_requested),
+            _json_array_or_none(record.target_names),
+            int(record.success),
+            _json_or_none(record.manifest),
+        ),
+    )
+    conn.commit()
 
 
 async def list_maintenance_runs(
